My Own (Entirely Self-built) Memory Game
I’ve decided to build a memory tile game from scratch as I wasn’t going
to take credit for the one we copied in class. It is definitely rough
around the edges (especially styling-wise) but that’s easy to fix. I’m
primarily proud of the logic (algorithm) behind it.